AndrewRichardA2Q4.js
This is the flag for Question 4. The design is derived from two countries' flags starting with my first and last initials, 'A' and 'R'. The countries in question are Angola and the Dominican *R*epublic.
I chose the Dominican Republic because "Republic" starts with R, and I did not like the designs or colors of the three nations that actually start with R (making a communist flag design with Russian colors seems distasteful, given the current geopolitical situation. Romania and Rwanda just have boring flags).
The Angolan color scheme combined with the Dominican Republic's cross gave me the impression that it could be a Scandinavian flag, so I designed the emblem in the center to have a Viking axe rather than a machete. I kept the star because communist nations love their star symbols, and the Angolan half-cogwheel gives the idea of an industrial people.
The final flag evokes the ideas of communism (red squares, yellow hammer and sickle-like design and star), and the cross indicates that it belongs to a Scandinavian nation, which I am naming "Folksland" or "People's Land", reminiscent of how Germany uses the word "Volk" to describe things that belong to the people (Volkswagen, Volksturmgewehr, etc.). Conveniently, the word for "folk" in all Scandinavian languages is still "folk".
All shapes were written as functions.
https://editor.p5js.org/Furi0usGeorgeX/sketches/hWecBRtLm